Convert Apache module to C++; add HTTP/JSON libraries
In order to simplify the HTTP requests and JSON handling for the API,
this converts the module to C++. These could certainly be done with C
libraries (e.g., libcurl and json-c), but the syntax provisions of
modern C++ should make this work more familiar and convenient. There are
expressive, header-only libraries that are easy to add without adding
complexity to the build process.

There are two such libraries vendored in the include directory:

- JSON for Modern C++ v3.10.5 - https://github.com/nlohmann/json
- cpp-httplib v0.10.7 - https://github.com/yhirose/cpp-httplib

Converting to C++ is straighforward. The Makefile has the compiler and
C++ standard level set, as well the additional include directory. The
code itself declares a couple of items as `extern "C"` so the Apache
server can call our entry points. Then, when loading the module, there
are prerequisites of loading libstdc++ and libssl.

There is one "trick" used in the Dockerfile to enable building on amd64
and aarch64: there is a cpu_arch variable defined in the httpd.conf with
`gcc -dumpmachine`, which is used in the library load paths. The base
image (httpd) is multi-arch, so no options are required to build on x86
or ARM by default, but you can specify the platform in the compose file.