Add files and config for test site
This brings in a static test site for acceptance testing. There is only
one of the resources really configured: "lit-authn", which is currently
requiring HTTP basic authentication to simulate a single sign-on
restriction, which should intercept the request before the module and,
when the basic auth is presented, should give REMOTE_USER to process.
This is modeling an enforce/active authentication scenario.