Set up hybrid build with Make and Bazel
The module is intended to be thin, because it is difficult to get Google
Test to run in the right context with the APXS/make build tooling. The
rest of the API client is a standalone package built with Bazel.

It is somewhat unlikely that we will use the C++ client for anything
other than the module, but it is nonetheless nice to have it
free-standing.

The Makefile actually builds everything, including the client, for the
full DSO. This is mostly because it was surprisingly challenging to use
the bazel build artifacts. It could certainly be done, but the
introduction of Bazel was primarily to support consistent and quick unit
testing for the client.