Use health check to establish DB dependency
There is apparently a condition in GitHub Actions where the network
configuration is not applied until the compose services are actually up,
but using a health check defers startup of a dependent long enough for
it to be in place. This also solves the problem of needing to sleep in
the database setup service command.