Restructure test-site to be more explanatory
This expands the concepts that were implied by the previous symbolic
names (e.g., lit-authn-or-ip to user-or-network) and includes a
description of why a request for one of these "collections" is expected
to be allowed or denied.

This also adjusts the existing system tests to the new paths and puts in
a rudimentary way to supply the test site URL via the TEST_URL_BASE
environment variable.