Revert default UID to 1000 and allow .env override
Docker Compose silently uses .env if it is present. This change passes
along a UID or GID in the Compose environment (via .env or arguments,
NOT everything from your shell environment) to the build args of the
Dockerfiles. All of the defaults are the usual 1000, but you can use
whatever you need (Linux users should likely use their own UID/GID, and
this likely applies in GitHub Actions as well).