Set up Hanami app Dockerfile/compose
This sets up a multi-stage build for the Hanami app, with production and
development gems/packages installed in separate images. The compose file
does not use the production image -- the development image is used for
both a webapp service and a "dev" service, intended to be used for
running tests and doing any bundler commands or the like.

- Remove api and cli services from compose file
- Add app (webapp) and app-dev (rspec/shell) services
- Add binstubs for our usual binaries
- Put bin/ on the PATH
- Install NodeJS from NodeSource repository