Add ROM boilerplate for database connections
We will need to reconcile the common notion of separate dev/test
databases with our earlier setup scripts, but this sets up the new
Hanami app to connect to "the database" via ROM and mysql2.

We can refer to the relations that are still under the top level lib
directory as we begin building out the API and need these tables mapped.