Localize mock server dependency to HttpClient test
The mock server is not intended to serve requests/responses for all
tests. We only intend to use it to ensure that we handle URLs, query
parameters, headers, and general HTTP matters correctly.

Everything "outward" from here should use mocks, generally, out to the
system tests, which will use a real API. The HttpClient response bodies
can be mocked inline in the ApiClient tests for locality and ease. The
mock really server only needs to run for http_client_test.cpp.