Add app healthcheck and dependency from tests
The start_interval/start_period are not working as intended -- there are
tracking issues in upstream moby, scheduled for 2.25 release. This means
that we have to bring the main interval down for it to be useful in
starting dependent services for run containers, because the first check
is delayed by the main interval.

This also adds the dependency so running system tests starts the app.
This should fix CI runs.