Adjust namespaces for ROM and domain objects
We started with our entities and repositories under the "persistence"
directory, but they are not really about persistence. The entities are
moved up to the top level in this commit, e.g., Lauth::Collection.

The repositories are moved up one level, retaining a collecting
namespace for them, since we have not yet decided whether to include a
suffix in their names. For example, we could have a
Lauth::Repositories::Grant or a Lauth::GrantRepo. Each of the choices
has subtle implications for filenames, verbosity, and visual parsing. So
we defer the final choice for now.