Find grants matching the request URI
There is subtle bug or missing feature of ROM-SQL with nested joins that
causes us to use a workaround for now. The Sequel::Dataset that each
relation defines fully supports multiple joins, but the join on
ROM::Relation only deals with the table for that relation.

In our case, we want grants JOIN collections JOIN locations, and however
we grouped it with the relations, there was a problem. Specifically,
these two fail differently:

- grants.join(collections.join(locations))
- grants.join(collections).join(locations)

So, for our workaround, we complete the join (and conditions on the
complete column set) at the Sequel layer, and then wrap the root
relation (Grants) around that dataset. This "wrapping" is what happens
in the rom-sql implementation of Relation#join, so we are not doing
anything too strange. It is just that, for convenience, Relation#join
determines the join keys by the defined associations. When doing so, it
joins to the table name of the right-hand relation rather than the
dataset. We can probably work with the ROM team to arrive at a patch,
since our desired behavior appears consistent with the design intent.