Call Ops::Authorize from Action::Authorize
This moves the access determination from the action to the authorize
operation, taking it out of the web context entirely. This gives us a
place to enrich the authorization logic in an isolated fashion.

The access request/result structs are rather anemic now, but they serve
the basic purpose. The Request should possibly get a factory method like
from_http_params that extracts what it needs from request.params. The
Result should likely get a to_json method, probably simply converting to
hash first.

We should also make types for allowed and denied results -- these are
primary concepts and ideal for inheritance. Whatever additional data we
need in the result can be in the base class with the specialization
being a const/literal return of "allowed" or "denied" from the
determination accessor.