Integrate individual and institution member handling
This uses a pretty straightforward left join approach. We join the users
for grants with userid set and the institution memberships for the
grants with inst set. Then we use a where condition with an OR on the
joined tables' userid to select to select where one of them matched.
They are mututally exclusive, so we don't get any undesired cross
products.