Add mod_authn_remoteuser
This standalone module allows configuring a location for header-based
authentication, including optional/anonymous.

It is activated with `AuthType RemoteUser`.

The header is configurable with `RemoteUserHeader Whatever-You-Want`.

When set as the AuthType, it always passes authentication. If the header
is not sent, it passes authentication while setting the anonymous
username. The username/REMOTE_USER is empty by default.

The anonymous username is configurable with
`RemoteUserAnonymousUsername whatever-you-want`.

The ap_auth_type is set to "RemoteUser".

There is also a debug CGI set up, using the printenv that is shipped
with the httpd image. It dumps the environment like we had in the
"delegated" script, but it uses env style syntax without the dependency
on the JSON::XS CPAN module.