Add option for dbsetup with keys but no fixtures
When creating a new database for sync, we want to set up the schema,
including the keys, but without any fixture data. This approach of
adding an option to setup.sh and the elif to add keys only is not ideal,
but need to leave any build automation working as-is. We should really
treat all of this and ensure that the logical and naming implications
are accurate and appropriate (e.g., "root" really means "create the
database and grant" and "all" really means "load fixture data").