Improve query performance by adjusting join conditions
We originally attempted to map everything out with left joins and then
apply the conditions that filter down to at most one matching row each
(for direct user grants, institution memberships, network, or group
memberships). It was conceptually correct, but the query optimizer did
not apply the WHERE fully through memberships joins. This meant that,
for institutions with many members, there was a cartesian product of
grants to memberships to process.

There are multiple approaches that solve this problem, including
adjusting the associativity/grouping of the joins (so that the
memberships inner join the users, thereby left joining at most one
membership-user row). However, the flat construction of the ROM query
code is attractive for its relative simplicity. By applying the literal
username comparison at the institution and group memberships, we achieve
the goal of left joining at most one membership.

In production-scale measurements, this amounts to the same results
being returned on the order of 2,000 times faster (e.g., 50ms vs 100s).