Export build artifacts for .deb packages
This uses shrink/actions-docker-extract and actions/upload-artifact to
take the completed .deb and .changes files out of a container that has
terminated, then upload them as a build artifact. This artifact is a
single zip file. We expect to eventually send a (GitHub) repository
event to apt-lib, which will trigger a workflow there to download the
build artifact, place the files in the appropriate incoming area, and
refresh the (APT) repository.