Add debug logging throughout API requests
This adds an indirect logging facility to the client and an
implementation that is injected from the module. It uses a globally
configured logger/singleton. The design is not perfect, but the
invocations and outputs are decent. I didn't want to add a logging
or formatting library at this time, but it's something to consider;
there are a few promising options. The C++20 format header is not
available under the current version of GCC, so we use basic
stringstreams, rather than dealing with varargs or adding {fmt}.