Ensure that SSL support is enabled in HttpClient
The http-check tool was working fine with HTTPS because it enabled the
support before it included httplib.h. However, the HttpClient
implementation did not, which mean that the whole of the shared library
was not supporting HTTPS.

When using the API over HTTPS, this was silently crashing the module
because of an uncaught exception. This would normally produce stdout or
stderr output, but it is squelched by the Apache process management.

We add an additional api-check binary that enables a cursory check of
using the Authorizer against an HTTP or HTTPS API instance from the
command line.