Add .rubocop.yml to ensure standard is used
For most cases, standard is recognized and used automatically, but in
some configurations, ruby-lsp was using plain rubocop. Since standard is
a simply a set of rubocop rules, rubocop can be configured to run with
them.

This should not be necessary, since ruby-lsp has a plugin/addon that is
supposed to be detected if standard is in your Gemfile... but, hey!