Fix conflict with delegation and authz type in system tests
There is an explicit conflict with delegation and pw auth: whether login
is required is now either an SSO config or app decision. With a
delegated resource, the expectation is that the app will decide on
authorization. The legacy behavior is that, when delegated and pw, to
pass entirely -- not computing the collections.

So, while conceptually valid, we do not want to test that combination
for A/B comparison. Switching the delegation tests to use "any" keeps
the behavior consistent and focused on delegation, rather than the edge
case in the legacy behavior.