Add group member system test
This is a simple change that exercises the basic function of group
membership. There is a new user, a group, a membership for the user in
the group, and a grant for the group to the by-username collection.

The group ID is not auto-incrementing, so we assign one that can be used
in production and stay out of the way.

The finer details of groups are fully tested at the API level; this is
for basic coverage and A/B testing.